fix(ci): address review feedback
- Share dashboard gRPC auth interceptor construction between deployments and realtime log streaming.
- Make the metadata error binding explicit for CodeQL analysis.

Refs: #847
