+++ dashboard/src/utils.rs
@@ -3,6 +3,8 @@
 //! Infrastructure utilities that don't follow the Django-style app structure.
 
 #[cfg(native)]
+pub(crate) mod grpc;
+#[cfg(native)]
 pub mod realtime;
 #[cfg(native)]
 pub mod vcs;
